examples/demo-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "remitmd-demo-agent"
version = "0.1.0"
description = "Autonomous demo agent for remit.md — runs all 4 payment lifecycle scenarios"
requires-python = ">=3.10"
dependencies = [
    "remitmd>=0.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
demo-agent = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
